broken: investigate type references for NumericMapping
In order to create NumericMapping, it's necessary to take a type that represents a mapping, extract the match type, and encode that match type in the mapping. A common example, specifying a type alias with a match type inside, doesn't work yet because getting the underlying type from a TypeRef seems difficult.