Shift direction on Mappings
Previously I was trying to construct a new Mapping based on the mapping given to my generator. This is not only difficult, but it complicates the implementation to needing a transparent modifier.

I've decided instead to check the mapping given and produce warnings or errors based on poorly defined mappings.